fix issue #76
