+++ http-sidecar/pom.xml
@@ -84,7 +84,22 @@
         </dependency>
         <dependency>
             <groupId>com.networknt</groupId>
+            <artifactId>egress-router</artifactId>
+            <version>${version.light-4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.networknt</groupId>
+            <artifactId>ingress-proxy</artifactId>
+            <version>${version.light-4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.networknt</groupId>
             <artifactId>health</artifactId>
+            <version>${version.light-4j}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.networknt</groupId>
+            <artifactId>http-url</artifactId>
             <version>${version.light-4j}</version>
         </dependency>
         <dependency>
@@ -196,7 +211,12 @@
             <version>${version.junit}</version>
             <scope>test</scope>
         </dependency>
-
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-inline</artifactId>
+            <version>3.8.0</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
