--- http-sidecar/src/main/java/com/networknt/router/HostWhitelist.java
@@ -1,44 +1,0 @@
-/*
- * Copyright (c) 2016 Network New Technologies Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.networknt.router;
-
-import com.networknt.config.Config;
-import com.networknt.config.ConfigException;
-
-import java.net.URI;
-import java.util.Arrays;
-
-public class HostWhitelist {
-
-    private static final RouterConfig config = (RouterConfig) Config.getInstance()
-            .getJsonObjectConfig("router", RouterConfig.class);
-
-    public boolean isHostAllowed(URI serviceUri) {
-        if (serviceUri != null) {
-            String[] hostWhitelist = config.getHostWhitelist();
-            if (hostWhitelist == null || hostWhitelist.length == 0) {
-                throw new ConfigException("No whitelist defined to allow the route to " + serviceUri);
-            }
-            String host = serviceUri.getHost();
-            return Arrays.stream(config.getHostWhitelist()).anyMatch(
-                    hostRegEx -> host != null && host.matches(hostRegEx));
-        } else {
-            return false;
-        }
-    }
-
-}
