+++ http-sidecar/src/test/java/com/networknt/proxy/SidecarProxyTest.java
@@ -122,6 +122,7 @@ public class SidecarProxyTest {
                 }
             });
 
+    @Ignore
     @Test
     public void testGet() throws Exception {
         final Http2Client client = Http2Client.getInstance();
@@ -159,6 +160,7 @@ public class SidecarProxyTest {
             Assert.assertTrue(reference.get().getAttachment(Http2Client.RESPONSE_BODY).contains("{\"backend\":\"OK\"}"));
         }
     }
+
     @Ignore
     @Test
     public void testPost() throws Exception {
@@ -220,7 +222,7 @@ public class SidecarProxyTest {
             }
             System.out.println(reference.get().getAttachment(Http2Client.RESPONSE_BODY));
             Assert.assertTrue(reference.get().getAttachment(Http2Client.RESPONSE_BODY).contains("{\"backend\":\"OK\"}"));
-            Thread.sleep(60000);
+            Thread.sleep(6000);
         }
     }
