+++ http-sidecar/src/test/java/com/networknt/router/middleware/ServiceDictHandlerTest.java
@@ -28,7 +28,7 @@ public class ServiceDictHandlerTest extends BaseServiceHandlerTest {
         if(server == null) {
             logger.info("starting server");
             HttpHandler handler = getTestHandler();
-            ServiceDictHandler serviceDictHandler = new ServiceDictHandler();
+            SidecarServiceDictHandler serviceDictHandler = new SidecarServiceDictHandler();
             serviceDictHandler.setNext(handler);
             handler = serviceDictHandler;
             HeaderHandler headerHandler = new HeaderHandler();
