+++ http-sidecar/src/test/resources/config/handler.yml
@@ -45,9 +45,9 @@ handlers:
   # Header middleware to manipulate request and/or response headers before or after downstream server
   - com.networknt.header.HeaderHandler@header
   # Direct requests to named services based on the request path
-  - com.networknt.router.middleware.SAMLTokenHandler@saml
-  - com.networknt.router.middleware.ServiceDictHandler@path
-  - com.networknt.router.RouterHandler@router
+  - com.networknt.router.middleware.SidecarSAMLTokenHandler@saml
+  - com.networknt.router.middleware.SidecarServiceDictHandler@path
+  - com.networknt.router.SidecarRouterHandler@router
   - com.networknt.proxy.LightProxyHandler@proxy
   - com.networknt.proxy.ProxyServerInfoHandler@proxyServerInfo
   # - com.networknt.resource.VirtualHostHandler@virtual
@@ -66,6 +66,7 @@ chains:
     - correlation
     # - cors
     - header
+   # - saml
     - path
     - router
     - specification
@@ -75,16 +76,6 @@ chains:
     #- sanitizer
     - validator
     - proxy
-  router:
-    - exception
-    # - metrics
-    - traceability
-    - correlation
-    # - cors
-    - header
-    # - saml
-    - path
-    - router
 
 paths:
   - path: '/*'
@@ -107,10 +98,6 @@ paths:
     method: 'patch'
     exec:
       - default
-  - path: '/v2/address'
-    method: 'get'
-    exec:
-      - router
 
   - path: '/health/com.networknt.backend-1.0.0'
     method: 'get'
